migrate entity detail
